Fix handling of default parameters in plans (#185)
Passing `None` as the default for a field sets the default to None and is
different to there being no default.

If a default value is given, it is now returned via a default_factory. This is
required to prevent pydantic errors when the default is an ophyd device. A
Callable wrapper class is used for the default_factory to allow the returned
value to be inspected (eg during testing)