Track completed statuses (#111)
For some devices there may be an additional status update broadcast
after the on_complete callback. This causes the status to be added
back to the dictionary where it will never be removed.

Track which statuses have been completed and check this set when
updating to ensure that a completed status is never added back to
the snapshot after on_complete is called.