Fix race condition on worker stop/start (#255)
The worker ensures that `worker.start()` blocks until the worker is
ready to execute a task and `worker.stop()` blocks until the worker is
no longer able to execute a task. It does so using a group of
`threading.Event`s.

There were intermittent issues when events were not being cleaned up
correctly, so when the worker was restarted it was not in the correct
state. This PR fixes those and cleans up some of the thread control
code.