Add page on adding plans and devices to blueapi (#291)
Document how to add plans and devices to blueapi.
This is an early draft meant to be criticised. The design it describes
is also not the final design and can definitely be improved upon. Would
be good to reduce the overall number of steps required to get a plan in!