Add support for scratch area to automatically add packages when running on Kubernetes (#313)
Recent discussion has emphasized the need for an area to checkout Python
packages and do dev work against a live instance of blueapi running on
Kubernetes. This has previously been done by pushing alpha tags of
various packages and continually rebuilding and redeploying the blueapi
container, which has become a lengthy and unweildy process.

This PR introduces the "scratch area", a shared directory that can be
mounted from the container running on kubernetes, and can install any
(pip installable) Python libraries that are cloned into it. In the cases
where a dev version of an existing dependency is required (e.g. Dodal),
the cloned version will overwrite the existing version installed in the
container.
Removing a package and restarting the container works like a factory
reset. See documentation addtions for more information.

Changes:
* Add confiugration for a scratch directory
* If configured, make blueapi pip install all packages in it on startup
* If configured, automatically create the scratch directory on startup
* Mount scratch in the helm chart if configured
* Tests and docs for new functionality