Make scratch area of startup script configurable for blueapi container (#327)
Currently, we have a linkam plan in i22-bluesky that makes an .xml file,
and then sets a PV value to retrieveing this xml.

Currently the blueapi container on p38 is failing to make this .xml file
because, although the correct location from Diamond workstations is
mounted into it, it's mounted in a different containerPath. This is a
problem, because it adds extra complexity to the bit of code in
i22-bluesky which needs to 1. create the xml file and 2. set the IOC to
look at it. The complexity being, that we now need to keep track of two
filepaths (one inside the container, one out).

Seeing as though we are currently reliant on the filesystem, it is not
much of a crime to demand that for now, the container scratch area is
configurable.