Add try except to messaging connection (#345)
Just a try except around the connection attempt. It displays the
failure to connect in the logs but does not fail fast. It allows the
rest of the application setup to complete just without a connection to a
message bus. You can then `crtl+c` to close it.

I was advised to just add this fix until the subprocess work is done
then we can re-address the behaviour we want.