Set umask in container
Performing an editable install on a shared filesystem scratch area
causes cache/version files to be written in the source repository.

With the container's default umask of 0022 the files are not group
writable which can lead to frustration when attempting to clean up the
scratch area.

Use the DLS "standard" of 0002.

Additionally modify the looping to only pickup directories within the
scratch area.

Fixes #478.