Remove handler
The handler has caused additional complexity with multiple levels of
indirection.

Remove the ABC and include the free functions in an "interface" module.

The main app and cli tests are rewritten to look more like unit tests.

This PR is to be followed up by further refactoring to replace the
singleton class with cached functions and restructuring the modules.

Fixes #434.