Use equality instead of identity to compare types (#533)
In some cases, possibly when MsgGenerator is defined or imported from
different modules, the types can be equivalent but are not the same
instance. In these cases, the identity check would fail and some plans
would be excluded where they shouldn't be. Using equality checks instead
mean it shouldn't matter where the MsgGenerator type was defined as all
instances should be equal.