Automatically init scratch area and remove startup script (#520)
Fixes #510 and #399 

Write python code to do the job of `container-startup.sh` and add code
to clone the required repos. Replace `container-startup.sh` with this
code. Stop blueapi from initializing scratch automatically on container
startup, replace this job with an
[initContainer](https://kubernetes.io/docs/concepts/workloads/pods/init-containers/).
Copy the blueapi venv to an
[emptyDir](https://kubernetes.io/docs/concepts/storage/volumes/#emptydir)
so the initContainers can manipulate it.

---------

Co-authored-by: Joe Shannon <joe.shannon@diamond.ac.uk>