Adding oauthn (#690)
### Alternative to the authentication were investigated:- 
[Client Libraries](https://oauth.net/code/python/) 

-  [Authomatic](http://authomatic.github.io/authomatic/)
- [ Python Social
Auth](https://python-social-auth.readthedocs.io/en/latest/) is an OAuth
and OAuth2 client for a multitude of services.
- [Authlib](https://github.com/lepture/authlib) has built-in OAuth 2
client for Flask and Django.
- [Requests-OAuthlib](https://github.com/requests/requests-oauthlib) has
OAuth library support for [Python
Requests](http://python-requests.org/).
- [mozilla-django-oidc](https://github.com/mozilla/mozilla-django-oidc)
is a lightweight Django authentication and access management library for
integration with OpenID Connect enabled authentication services.

In the end I decided to not use them as we just need to make 2 requests
and there is not that much error handling required as well ... We can
look into integrating on of the above mentioned alternatives for the
OAuth device flow integration.

---------

Co-authored-by: Ware, Joseph (DLSLtd,RAL,LSCI) <joseph.ware@diamond.ac.uk>
Co-authored-by: Joseph Ware <53935796+DiamondJoseph@users.noreply.github.com>