Add fake `PathProvider` to startup module (#783)
This is a workaround (not a fix) for #784, so won't close it. The
replacement of `PathProvider` in dodal will eventually provide the
proper fix. For now I think we should just pretend to write to a
directory with our start-up file, despite the fact that none of the
simulated detectors actually write files.