Use an external directory for container python cache
With this set, the pyc files are created in a mirror directory instead
of the source tree. When running in a shared directory (as is the case
for the scratch area used by bluapi), this prevents files being created
with permissions that prevent other users removing them.

See python docs at
https://docs.python.org/3/using/cmdline.html#envvar-PYTHONPYCACHEPREFIX