Use an external directory for container python cache (#817)
With this set, the pyc files are created in a mirror directory instead
of the source tree. When running in a shared directory (as is the case
for the scratch area used by bluapi), this prevents files being created
with permissions that prevent other users removing them.

See python docs at
https://docs.python.org/3/using/cmdline.html#envvar-PYTHONPYCACHEPREFIX

Fixes #812

### Instructions to reviewer on how to test:
1. Deploy and run on p46 with new scratch area
2. Check that there are no `__pycache__` directories created in the
scratch area

### Checks for reviewer
- [X] Would the PR title make sense to a user on a set of release notes