Always construct auth route (#889)
Allows the `/config/oidc` endpoint to return `204 No Content` instead of
a OIDCConfig object. This represents to the client that authentication
is not configured/not required, allowing the schema of routes to be
consistent whether config if configured or not.

This does not resolve the SecuritySchema being present/absent in the
schema.