add pydantic level validation for blueapi to not be a valid target (#893)
solves #723 

Added a check in scratch.py at the cloning stage (before even pip
install). it is kind of redundant - as I also added a check on the
pydantic validation of the config that happens earlier than that. The
former might be useful if someone chooses to provide config not-through
pydantic models - not sure how that would happen but it's more robust
that way. Also added a test for parsing the config.