fix:  change set up order of the worker to make a StartDocumentPathProvider before device instantiation (#941)
This fixes the order of configuration of the context. If numtracker is configured, a StartDocumentPathProvider is created. The devices,  which are then loaded from device module(s), will use that PathProvider when they are instantiated. If there is a PathProvider defined in the device module(s), an exception will be raised. It will need to be removed in order to use numtracker.