ci: Group and disable automated updates of dependencies that are updated by the copier template (#975)
Prevents the state where we have updated beyond the copier template and
it's unclear as the actions are pinned to commit SHA, so we revert only
to be updated again.
Will require checking: Are we using these actions outside of the files
generated by the copier template (e.g. in our Helm build?), possibly
re-generating our CI with the latest version of the template. Also
assumes that the copier template adopts automation