ci: Add CODEOWNERS (#1049)
Add CODEOWNERS of the DAQ core team, to allow opening the repository up
to general editing but require an approving code review from a
CODEOWNER, rather than allowing all members of the org to contribute and
review simultaneously. This is intended as a living document and is not
an end to discussion on codeownership, just a reflection of the current
state.