fix!: Remove ophyd (#1087)
Removes our direct dependency on ophyd. This was intended to fix some
spurious log messages on startup, but we still get ophyd as a dependency
of dodal. It at least allows us to only have that dependency via dodal,
and when dodal no longer requires ophyd it will not still be a trailing
edge.

This is a breaking change to anyone using config generated from the
default