fix!: Remove instrument_session from config (#1116)
Removes the instrument session from the config, which is no longer
required as of 0.19.0 requiring the instrument_session being passed with
the request as of #1062

This is a breaking change as our Config classes do not allow additional
fields. An alternative is to deprecate the field and default it to None,
but as this is still pre-1.0.0 I thought it was easier to bundle this
change in.