fix: Pin the system test containers (#1144)
The example-services compose-file that we for our system tests is
[adjusting the ports that it
uses](https://github.com/epics-containers/example-services/pull/43). As
we are tracking main of that repository we need to pin this dependency,
allowing the ports to be adjusted for the next release.

With Renovate we should be able to keep the dependency pinned and use
the latest releases, which will require an adjustment to the ports when
the release is made.