test: Add example services as a submodule (#1155)
Add example services as a submodule and include it in the system tests
docker compose file, rewrite the system test docs to make use of this.

The system tests will now be easier to run and debug locally, since this
automates a lot of the setup and tear down process.