Ensure that numtracker headers are not re-used
If the headers are only set when they are present but not cleared at the
end of a plan (or when they're absent), a request with no headers
following a request with headers will re-use the previous users
credentials allowing unauthenticated access to running plans.

Setting the headers to an empty map if none are present helps ensure
tokens cannot be used by other users.