fix(helm): Run Dockerfile as uid 1000 (#1274)
Configures a default 1000 user for container to run as. This allows a
username to be found when running without the LDAP sidecar or when
running as a user not registered in LDAP, e.g. the default 1000 user in
the Helm chart/Kyverno policy.

As we remove the need for hostNetwork and continue to push for a
non-dls_sw scratch area, the container does not require running as a
specific user and can run as an arbitrary user with fewer permissions.

This fixes problems of [attaching a devcontainer for
debugging](https://diamondlightsource.github.io/python-copier-template/main/how-to/debug-in-cluster.html#debugging-in-the-cluster)
when running as a non-LDAP user.