feat: Seperate debug and suspend behaviours; consolidate LDAP requirements (#1130)
Refactors the container build to remove the debug container, and
properly capture the requirements on the LDAP sidecar.

This container build has been used to debug using VSCode remote
attaching to the container, or by port-forwarding from a local
workstation.

The previous `debug` behaviour of suspending has been extracted into a
`suspend` argument.