Unsubscribe the callback that unsubscribes the tiled writer
When the tiled writer subscribes to the run_engine, another callback
subscribes to the task worker to remove the writer at the end of the
plan/task. This subscriber should also be removed at the end of the
task.

To avoid creating a third subscriber and the same issue again, rely on
the unsubscribe handle created after the callback is defined being
available in the closure when it runs. There is a very slight race
condition here if the task complete event is received before the
subscribe method returns but until it happens we'll assume it's
unlikely.